refactor: Refactor login_manager.py, pages.py, auth_routes.py, app.py, contact.py, and dash.html files
