+++ requirements.txt
@@ -5,6 +5,7 @@ charset-normalizer==3.3.2
 click==8.1.7
 cryptography==42.0.5
 Flask==3.0.3
+Flask-Login==0.6.3
 greenlet==3.0.3
 gunicorn==21.2.0
 idna==3.7
