+++ src/templates/dashboard.html
@@ -16,10 +16,7 @@
                         <h2>Clientes</h2>
                     </div>
                     <div class="card-body">
-                        <a href="https://roines-customer-manager.onrender.com/api/customers" class="btn btn-primary mb-2">Visualizar Clientes</a>
-                        <a href="#" class="btn btn-success mb-2">Adicionar Cliente</a>
-                        <a href="#" class="btn btn-warning mb-2">Editar Cliente</a>
-                        <a href="#" class="btn btn-danger">Excluir Cliente</a>
+                        <a href="/customers" class="btn btn-primary mb-2">Gerenciar Clientes</a>
                     </div>
                 </div>
             </div>
@@ -29,11 +26,8 @@
                         <h2>Contatos</h2>
                     </div>
                     <div class="card-body">
-
-                        <a href="https://roines-customer-manager.onrender.com/api/contacts" class="btn btn-primary mb-2">Visualizar Contatos</a>
-                        <a href="#" class="btn btn-success mb-2">Adicionar Contato</a>
-                        <a href="#" class="btn btn-warning mb-2">Editar Contato</a>
-                        <a href="#" class="btn btn-danger">Excluir Contato</a>
+                        <!-- Link para a página de gerenciamento de contatos (ainda não desenvolvida) -->
+                        <a href="/contacts" class="btn btn-primary mb-2">Gerenciar Contatos</a>
                     </div>
                 </div>
             </div>
@@ -62,11 +56,9 @@
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.4/dist/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 
-
     <!-- JavaScript para preencher o relatório -->
     <script>
-       
-        fetch('https://roines-customer-manager.onrender.com/api/customers/report')
+        fetch('/api/report')
         .then(response => response.json())
         .then(data => fillReport(data))
