+++ src/templates/login.html
@@ -92,7 +92,7 @@
     </div>
     <div class="card-body text-center">
       <p>Faça login com sua conta do Google</p>
-      <a href="https://roines-customer-manager.onrender.com/auth/login" class="btn btn-lg btn-google btn-block">
+      <a href="/auth/login" class="btn btn-lg btn-google btn-block">
         <i class="fab fa-google mr-2"></i> Entrar com Google
       </a>
     </div>
